Add ability to patch RSTB via commandline
When more than one argument is supplied for RSTB editing, the program will assume the 2nd argument is the path to the romFS folder and use that to overwrite the input RSTB with a patched one.
If a third argument is supplied, it will assume the 3rd path is the output RSTB and save it there instead of overwriting.
To make this work without further user input, I had to add bools for whether or not to show messagebox prompts to the RSTB saving and updating functions. This way, the program assumes that you always want to compress the RSTB and doesn't show the results messagebox when updating via commandline.

One issue I've noticed is that occasionally there is a popup about the collection being modified while in a for loop. I assume this is some kind of race condition relating to updating being an async function, but I'm not sure.